This is Lab 4 - From CSC110
This program asks the user to enter a number of seconds. The program then calculates how many days, hours, minutes, and seconds are in the entered number. Next, the program outputs the number of hours, minutes, and seconds that corresponds to the total seconds entered.
Finally at the end the program asks the user if they want to continue or exit.